Renamet gamle behandlingsmenyer til legacybehandlingsmeny
