+++ src/sider/eu_eøs/saksbehandling/komponenter/legacybehandlingsmeny.js
@@ -3,9 +3,9 @@ import PT from "prop-types";
 
 import * as Nav from "../../../../utils/navFrontend";
 
-import "./behandlingsmeny.css";
+import "./legacybehandlingsmeny.css";
 
-const Behandlingsmeny = ({
+const Legacybehandlingsmeny = ({
   redigerbart,
   lagreOgLukkHandle,
   tilbakeleggeHandle,
@@ -13,8 +13,10 @@ const Behandlingsmeny = ({
   apneTidligereBehandlinger,
   visAvsluttSakSomBortfaltDialogHandle,
   visHenleggSak,
-  visAvslagManglendeOpplysninger,
   visAvslagSoknadDialogHandle,
+  visAvslagManglendeOpplysninger,
+  visRevurderFagsakDialogHandle,
+  visRevurderFagsak,
 }) => (
   <Nav.EkspanderbartpanelBase
     ariaTittel="Behandlingsmeny"
@@ -48,11 +50,16 @@ const Behandlingsmeny = ({
       <Nav.Knapp mini className="innhold__element" onClick={apneTidligereBehandlinger}>
         Vis alle behandlinger
       </Nav.Knapp>
+      {visRevurderFagsak && (
+        <Nav.Knapp mini className="innhold__element" onClick={visRevurderFagsakDialogHandle}>
+          Vurder saken på nytt
+        </Nav.Knapp>
+      )}
     </div>
   </Nav.EkspanderbartpanelBase>
 );
 
-Behandlingsmeny.propTypes = {
+Legacybehandlingsmeny.propTypes = {
   lagreOgLukkHandle: PT.func.isRequired,
   tilbakeleggeHandle: PT.func.isRequired,
   visHenleggDialogHandle: PT.func.isRequired,
@@ -61,7 +68,9 @@ Behandlingsmeny.propTypes = {
   apneTidligereBehandlinger: PT.func.isRequired,
   redigerbart: PT.bool.isRequired,
   visHenleggSak: PT.bool.isRequired,
+  visRevurderFagsakDialogHandle: PT.func.isRequired,
   visAvslagManglendeOpplysninger: PT.bool.isRequired,
+  visRevurderFagsak: PT.bool.isRequired,
 };
 
-export default Behandlingsmeny;
+export default Legacybehandlingsmeny;
