+++ src/sider/ftrl/saksbehandling/legacybehandlingsmeny.less
@@ -1,7 +1,7 @@
-@import "../../utils/mixins";
-@import "../../constants";
+@import "../../../utils/mixins";
+@import "../../../constants";
 
-.behandlingsmeny {
+.legacybehandlingsmeny {
   position: absolute;
   z-index: 2;
   background-color: @white;
