--- src/sider/rammeverk/komponenter/personlinje/behandlingsmeny.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import * as Ikon from "../../../../resources/images";
-import "./personlinje.css";
-
-const Behandlingsmeny = () => {
-  return (
-    <div className="personlinje__behandlingsmeny">
-      <Ikon.Hamburger className="hamburger" />
-    </div>
-  );
-};
-
-export default Behandlingsmeny;
